feature : OrderControllerTest / refactor : entity -> item, order

+++ backend/project1/src/main/java/com/team4/project1/domain/item/controller/ApiV1ItemController.java
@@ -26,7 +26,7 @@ public class ApiV1ItemController {
 
 
     @GetMapping("/{itemId}")
-    public ItemDto item(@PathVariable Integer itemId) {
+    public ItemDto item(@PathVariable Long itemId) {
         return itemService.getItemById(itemId).orElseThrow(() -> new ItemNotFoundException(itemId));
     }
 }
