+++ backend/project1/src/main/java/com/team4/project1/domain/item/repository/ItemRepository.java
@@ -2,13 +2,12 @@ package com.team4.project1.domain.item.repository;
 
 import com.team4.project1.domain.item.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
     List<Item> findAllByOrderByPriceAsc();
     List<Item> findAllByOrderByNameAsc();
 }
